refactor(ui): extract Details tab to ui/tabs/details.rs
Lifts draw_connection_details (~1,100 lines) and its six
push_detail/register_detail helpers into ui/tabs/details.rs
(~1,225 lines). With this, all five tabs live under ui/tabs/.

mod.rs is now ~990 lines (from the original 5,100): dispatcher,
panel_block, the four cross-tab cell-style helpers, NO_COLOR
plumbing, the two DETAIL constants, and the snapshot tests
module — and nothing tab-specific.
