+++ src/ui/tabs/mod.rs
@@ -2,6 +2,7 @@
 //! tab (Overview, Details, Interfaces, Graph, Help) and is invoked
 //! from the top-level `draw()` dispatcher in `ui::mod`.
 
+pub(super) mod details;
 pub(super) mod graph;
 pub(super) mod help;
 pub(super) mod interfaces;
